% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSTEM.R
\name{run_scstem_GUI}
\alias{run_scstem_GUI}
\title{scSTEM GUI function}
\usage{
run_scstem_GUI()
}
\description{
This function runs scSTEM GUI
}
\details{
To use scSTEM in R, Run \code{run_scstem_GUI()} in R to call a shiny app based GUI. Please follow the steps described in user manual to perform scSTEM analysis.
}
